requirements/requirements.txt
"""
Archivo para saber las librerías instaladas
cmd=python
cmd=python -m pip freeze
PostgreSQL 12.2
pgAdmin 4 4.18
"""
asgiref==3.2.7
Django==3.0.3
django-crum==0.7.6
django-widget-tweaks==1.4.8
entrypoints==0.3
flake8==3.7.9
html5lib==1.1
mccabe==0.6.1
Pillow==7.2.0
psycopg2==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
pyPDF2==1.26.0
python==3.8.2
pytz==2019.3
reportlab==3.5.46
six==1.15.0
sqlparse==0.3.1
webencodings==0.5.1
xhtml2pdf==0.2.4
